new-fastapi/posts/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.5.0
ecdsa==0.18.0
fastapi==0.108.0
fastapi_auth_middleware==1.0.2
fastapi_keycloak==1.0.10
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.32.0.post1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
pymongo==4.6.1
pika==1.3.2
py-eureka-client==0.11.9
httpx==0.23.2
fastapi-responses==0.2.1
xmltodict==0.13.0
aio-pika==9.4.0 